refactor: navigation
